refactor(sagemaker): consolidate SageMaker types into centralized provider types
Moves all SageMaker-specific type definitions from standalone types.ts module
into the centralized src/lib/types/providers.ts architecture.

## Changes
- **Moved types**: All SageMaker types (489 lines) from sagemaker/types.ts to types/providers.ts
- **Updated imports**: All SageMaker modules now import from "../../types/providers.js"
- **Deleted file**: Removed standalone src/lib/providers/sagemaker/types.ts
- **Centralized architecture**: SageMaker types now part of unified provider type system

## Benefits
- Consistent with overall types module architecture
- Single source of truth for all provider types
- Improved maintainability and discoverability
- Aligned with centralized type management strategy
